+++ Libraries/Lists/FlatList.js.flow
@@ -10,7 +10,6 @@
 
 const React = require('react');
 const View = require('../Components/View/View');
-const VirtualizedList = require('./VirtualizedList');
 
 import typeof ScrollViewNativeComponent from '../Components/ScrollView/ScrollViewNativeComponent';
 import {type ScrollResponderType} from '../Components/ScrollView/ScrollView';
@@ -20,6 +19,7 @@ import type {
   ViewabilityConfigCallbackPair,
 } from './ViewabilityHelper';
 import type {RenderItemType, RenderItemProps} from './VirtualizedList';
+import typeof VirtualizedList from './VirtualizedList';
 
 type RequiredProps<ItemT> = {|
   /**
@@ -61,7 +61,7 @@ type FlatListProps<ItemT> = {|
   ...OptionalProps<ItemT>,
 |};
 
-type VirtualizedListProps = React.ElementConfig<typeof VirtualizedList>;
+type VirtualizedListProps = React.ElementConfig<VirtualizedList>;
 
 export type Props<ItemT> = {
   ...$Diff<
