+++ Libraries/Lists/__tests__/VirtualizedList-test.js
@@ -10,18 +10,9 @@
 
 'use strict';
 
-import VirtualizedList_EXPERIMENTAL from '../VirtualizedList_EXPERIMENTAL';
-import * as VirtualizedListInjection from '../VirtualizedListInjection';
+import VirtualizedList from '../VirtualizedList';
 import React from 'react';
 import ReactTestRenderer from 'react-test-renderer';
-
-const VirtualizedList = require('../VirtualizedList');
-
-const useExperimentalList =
-  process.env.USE_EXPERIMENTAL_VIRTUALIZEDLIST === 'true';
-if (useExperimentalList) {
-  VirtualizedListInjection.inject(VirtualizedList_EXPERIMENTAL);
-}
 
 describe('VirtualizedList', () => {
   it('renders simple list', () => {
@@ -785,12 +776,7 @@ it('scrolls after content sizing with fractional initialScrollIndex (getItemLayo
     performAllBatches();
   });
 
-  if (useExperimentalList) {
-    expect(scrollTo).toHaveBeenLastCalledWith({y: 2.0, animated: false});
-  } else {
-    // Legacy incorrect results
-    expect(scrollTo).toHaveBeenLastCalledWith({y: Number.NaN, animated: false});
-  }
+  expect(scrollTo).toHaveBeenLastCalledWith({y: 2.0, animated: false});
 });
 
 it('scrolls after content sizing with fractional initialScrollIndex (cached layout)', () => {
@@ -828,12 +814,7 @@ it('scrolls after content sizing with fractional initialScrollIndex (cached layo
     performAllBatches();
   });
 
-  if (useExperimentalList) {
-    expect(scrollTo).toHaveBeenLastCalledWith({y: 2.0, animated: false});
-  } else {
-    // Legacy incorrect results
-    expect(scrollTo).toHaveBeenLastCalledWith({y: 8.25, animated: false});
-  }
+  expect(scrollTo).toHaveBeenLastCalledWith({y: 2.0, animated: false});
 });
 
 it('scrolls after content sizing with fractional initialScrollIndex (layout estimation)', () => {
@@ -1754,187 +1735,183 @@ it('calls _onCellLayout properly', () => {
   expect(mock).not.toHaveBeenCalledWith(event, 'i3', 2);
 });
 
-if (useExperimentalList) {
-  describe('VirtualizedList (Experimental functionality)', () => {
-    it('keeps viewport below last focused rendered', () => {
-      const items = generateItems(20);
-      const ITEM_HEIGHT = 10;
+it('keeps viewport below last focused rendered', () => {
+  const items = generateItems(20);
+  const ITEM_HEIGHT = 10;
 
-      let component;
-      ReactTestRenderer.act(() => {
-        component = ReactTestRenderer.create(
-          <VirtualizedList
-            initialNumToRender={1}
-            windowSize={1}
-            {...baseItemProps(items)}
-            {...fixedHeightItemLayoutProps(ITEM_HEIGHT)}
-          />,
-        );
-      });
-
-      ReactTestRenderer.act(() => {
-        simulateLayout(component, {
-          viewport: {width: 10, height: 50},
-          content: {width: 10, height: 200},
-        });
-
-        performAllBatches();
-      });
-
-      ReactTestRenderer.act(() => {
-        component.getInstance()._onCellFocusCapture(3);
-      });
-
-      ReactTestRenderer.act(() => {
-        simulateScroll(component, {x: 0, y: 150});
-        performAllBatches();
-      });
-
-      // Cells 1-8 should remain rendered after scrolling to the bottom of the list
-      expect(component).toMatchSnapshot();
-    });
-
-    it('virtualizes away last focused item if focus changes to a new cell', () => {
-      const items = generateItems(20);
-      const ITEM_HEIGHT = 10;
-
-      let component;
-      ReactTestRenderer.act(() => {
-        component = ReactTestRenderer.create(
-          <VirtualizedList
-            initialNumToRender={1}
-            windowSize={1}
-            {...baseItemProps(items)}
-            {...fixedHeightItemLayoutProps(ITEM_HEIGHT)}
-          />,
-        );
-      });
-
-      ReactTestRenderer.act(() => {
-        simulateLayout(component, {
-          viewport: {width: 10, height: 50},
-          content: {width: 10, height: 200},
-        });
-
-        performAllBatches();
-      });
-
-      ReactTestRenderer.act(() => {
-        component.getInstance()._onCellFocusCapture(3);
-      });
-
-      ReactTestRenderer.act(() => {
-        simulateScroll(component, {x: 0, y: 150});
-        performAllBatches();
-      });
-
-      ReactTestRenderer.act(() => {
-        component.getInstance()._onCellFocusCapture(17);
-      });
-
-      // Cells 1-8 should no longer be rendered after focus is moved to the end of
-      // the list
-      expect(component).toMatchSnapshot();
-    });
-
-    it('keeps viewport above last focused rendered', () => {
-      const items = generateItems(20);
-      const ITEM_HEIGHT = 10;
-
-      let component;
-      ReactTestRenderer.act(() => {
-        component = ReactTestRenderer.create(
-          <VirtualizedList
-            initialNumToRender={1}
-            windowSize={1}
-            {...baseItemProps(items)}
-            {...fixedHeightItemLayoutProps(ITEM_HEIGHT)}
-          />,
-        );
-      });
-
-      ReactTestRenderer.act(() => {
-        simulateLayout(component, {
-          viewport: {width: 10, height: 50},
-          content: {width: 10, height: 200},
-        });
-
-        performAllBatches();
-      });
-
-      ReactTestRenderer.act(() => {
-        component.getInstance()._onCellFocusCapture(3);
-      });
-
-      ReactTestRenderer.act(() => {
-        simulateScroll(component, {x: 0, y: 150});
-        performAllBatches();
-      });
-
-      ReactTestRenderer.act(() => {
-        component.getInstance()._onCellFocusCapture(17);
-      });
-
-      ReactTestRenderer.act(() => {
-        simulateScroll(component, {x: 0, y: 0});
-        performAllBatches();
-      });
-
-      // Cells 12-19 should remain rendered after scrolling to the top of the list
-      expect(component).toMatchSnapshot();
-    });
-
-    it('virtualizes away last focused index if item removed', () => {
-      const items = generateItems(20);
-      const ITEM_HEIGHT = 10;
-
-      let component;
-      ReactTestRenderer.act(() => {
-        component = ReactTestRenderer.create(
-          <VirtualizedList
-            initialNumToRender={1}
-            windowSize={1}
-            {...baseItemProps(items)}
-            {...fixedHeightItemLayoutProps(ITEM_HEIGHT)}
-          />,
-        );
-      });
-
-      ReactTestRenderer.act(() => {
-        simulateLayout(component, {
-          viewport: {width: 10, height: 50},
-          content: {width: 10, height: 200},
-        });
-
-        performAllBatches();
-      });
-
-      ReactTestRenderer.act(() => {
-        component.getInstance()._onCellFocusCapture(3);
-      });
-
-      ReactTestRenderer.act(() => {
-        simulateScroll(component, {x: 0, y: 150});
-        performAllBatches();
-      });
-
-      const itemsWithoutFocused = [...items.slice(0, 3), ...items.slice(4)];
-      ReactTestRenderer.act(() => {
-        component.update(
-          <VirtualizedList
-            initialNumToRender={1}
-            windowSize={1}
-            {...baseItemProps(itemsWithoutFocused)}
-            {...fixedHeightItemLayoutProps(ITEM_HEIGHT)}
-          />,
-        );
-      });
-
-      // Cells 1-8 should no longer be rendered
-      expect(component).toMatchSnapshot();
-    });
+  let component;
+  ReactTestRenderer.act(() => {
+    component = ReactTestRenderer.create(
+      <VirtualizedList
+        initialNumToRender={1}
+        windowSize={1}
+        {...baseItemProps(items)}
+        {...fixedHeightItemLayoutProps(ITEM_HEIGHT)}
+      />,
+    );
   });
-}
+
+  ReactTestRenderer.act(() => {
+    simulateLayout(component, {
+      viewport: {width: 10, height: 50},
+      content: {width: 10, height: 200},
+    });
+
+    performAllBatches();
+  });
+
+  ReactTestRenderer.act(() => {
+    component.getInstance()._onCellFocusCapture(3);
+  });
+
+  ReactTestRenderer.act(() => {
+    simulateScroll(component, {x: 0, y: 150});
+    performAllBatches();
+  });
+
+  // Cells 1-8 should remain rendered after scrolling to the bottom of the list
+  expect(component).toMatchSnapshot();
+});
+
+it('virtualizes away last focused item if focus changes to a new cell', () => {
+  const items = generateItems(20);
+  const ITEM_HEIGHT = 10;
+
+  let component;
+  ReactTestRenderer.act(() => {
+    component = ReactTestRenderer.create(
+      <VirtualizedList
+        initialNumToRender={1}
+        windowSize={1}
+        {...baseItemProps(items)}
+        {...fixedHeightItemLayoutProps(ITEM_HEIGHT)}
+      />,
+    );
+  });
+
+  ReactTestRenderer.act(() => {
+    simulateLayout(component, {
+      viewport: {width: 10, height: 50},
+      content: {width: 10, height: 200},
+    });
+
+    performAllBatches();
+  });
+
+  ReactTestRenderer.act(() => {
+    component.getInstance()._onCellFocusCapture(3);
+  });
+
+  ReactTestRenderer.act(() => {
+    simulateScroll(component, {x: 0, y: 150});
+    performAllBatches();
+  });
+
+  ReactTestRenderer.act(() => {
+    component.getInstance()._onCellFocusCapture(17);
+  });
+
+  // Cells 1-8 should no longer be rendered after focus is moved to the end of
+  // the list
+  expect(component).toMatchSnapshot();
+});
+
+it('keeps viewport above last focused rendered', () => {
+  const items = generateItems(20);
+  const ITEM_HEIGHT = 10;
+
+  let component;
+  ReactTestRenderer.act(() => {
+    component = ReactTestRenderer.create(
+      <VirtualizedList
+        initialNumToRender={1}
+        windowSize={1}
+        {...baseItemProps(items)}
+        {...fixedHeightItemLayoutProps(ITEM_HEIGHT)}
+      />,
+    );
+  });
+
+  ReactTestRenderer.act(() => {
+    simulateLayout(component, {
+      viewport: {width: 10, height: 50},
+      content: {width: 10, height: 200},
+    });
+
+    performAllBatches();
+  });
+
+  ReactTestRenderer.act(() => {
+    component.getInstance()._onCellFocusCapture(3);
+  });
+
+  ReactTestRenderer.act(() => {
+    simulateScroll(component, {x: 0, y: 150});
+    performAllBatches();
+  });
+
+  ReactTestRenderer.act(() => {
+    component.getInstance()._onCellFocusCapture(17);
+  });
+
+  ReactTestRenderer.act(() => {
+    simulateScroll(component, {x: 0, y: 0});
+    performAllBatches();
+  });
+
+  // Cells 12-19 should remain rendered after scrolling to the top of the list
+  expect(component).toMatchSnapshot();
+});
+
+it('virtualizes away last focused index if item removed', () => {
+  const items = generateItems(20);
+  const ITEM_HEIGHT = 10;
+
+  let component;
+  ReactTestRenderer.act(() => {
+    component = ReactTestRenderer.create(
+      <VirtualizedList
+        initialNumToRender={1}
+        windowSize={1}
+        {...baseItemProps(items)}
+        {...fixedHeightItemLayoutProps(ITEM_HEIGHT)}
+      />,
+    );
+  });
+
+  ReactTestRenderer.act(() => {
+    simulateLayout(component, {
+      viewport: {width: 10, height: 50},
+      content: {width: 10, height: 200},
+    });
+
+    performAllBatches();
+  });
+
+  ReactTestRenderer.act(() => {
+    component.getInstance()._onCellFocusCapture(3);
+  });
+
+  ReactTestRenderer.act(() => {
+    simulateScroll(component, {x: 0, y: 150});
+    performAllBatches();
+  });
+
+  const itemsWithoutFocused = [...items.slice(0, 3), ...items.slice(4)];
+  ReactTestRenderer.act(() => {
+    component.update(
+      <VirtualizedList
+        initialNumToRender={1}
+        windowSize={1}
+        {...baseItemProps(itemsWithoutFocused)}
+        {...fixedHeightItemLayoutProps(ITEM_HEIGHT)}
+      />,
+    );
+  });
+
+  // Cells 1-8 should no longer be rendered
+  expect(component).toMatchSnapshot();
+});
 
 function generateItems(count) {
   return Array(count)
