--- Libraries/Lists/__tests__/VirtualizedList_EXPERIMENTAL-test.js
@@ -1,14 +1,0 @@
-/**
- * Copyright (c) Meta Platforms, Inc. and affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- *
- * @format
- * @oncall react_native
- */
-
-'use strict';
-
-process.env.USE_EXPERIMENTAL_VIRTUALIZEDLIST = 'true';
-require('./VirtualizedList-test');
